requirements.txt
pymongo
flask
Flask-Cors
scikit-learn
newsapi-python
# difflib